backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
google-auth==2.36.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
jwt==1.3.1
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
